Fix AVX-512 intrinsics header for GCC 7.3
Ugly hack so we can build OpenBLAS using the older gcc release (7.3.0)
that's the linux-64 compiler for default; without this, compilation of
Skylake X kernels (and possibly others) will fail with `incompatible
type for argument 1 of '_mm512_abs_pd'` errors.

Using this hack because updating the compilers for the "defaults"
channel is much more complicated than we had anticipated.